requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
pydantic==2.11.7
openai==1.101.0
python-dotenv==1.1.1
